+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to Semantic Versioning(http://semver.org/).
 
+## 2.6.1
+- Change order and product tagging to header / footer
 
 ## 2.6.0
 - Add support for multiple currencies (using exchange rates)